javascript objects intro - create object and apply properties and a method to alert say hello
Please follow the instructions below.

Create an HTML file named objects-intro-js.html within the
~/vagrant-lamp/sites/codeup.dev/public folder on your Mac.
Copy and paste the code below as inline JavaScript and complete the
todo items in the code.
View the results in your browser to test for the expected output.
Finally, save your work, commit the changes to your git repository, and
push to github.